put SettingsParser in a separate file and add test for --s-from-file
